Add single_result.hpp content to helpers.hpp
