+++ libs/pika/execution/include/pika/execution/algorithms/sync_wait.hpp
@@ -14,7 +14,7 @@
 
 #include <pika/concepts/concepts.hpp>
 #include <pika/datastructures/variant.hpp>
-#include <pika/execution/algorithms/detail/single_result.hpp>
+#include <pika/execution/algorithms/detail/helpers.hpp>
 #include <pika/execution_base/operation_state.hpp>
 #include <pika/execution_base/receiver.hpp>
 #include <pika/execution_base/sender.hpp>
