+++ libs/pika/execution/include/pika/execution/algorithms/when_all.hpp
@@ -14,7 +14,7 @@
 #include <pika/concepts/concepts.hpp>
 #include <pika/datastructures/member_pack.hpp>
 #include <pika/datastructures/variant.hpp>
-#include <pika/execution/algorithms/detail/single_result.hpp>
+#include <pika/execution/algorithms/detail/helpers.hpp>
 #include <pika/execution_base/operation_state.hpp>
 #include <pika/execution_base/receiver.hpp>
 #include <pika/execution_base/sender.hpp>
